employee.js가 너무 많은 기능을 포함하고 있어 format, resignation, file로 분리와 js 디렉토리 정리
