--- src/main/java/com/woosan/hr_system/schedule/controller/ScheduleController.java
@@ -1,4 +1,0 @@
-package com.woosan.hr_system.schedule.controller;
-
-public class ScheduleController {
-}
