--- src/main/java/com/woosan/hr_system/schedule/dao/ClientDAO.java
@@ -1,4 +1,0 @@
-package com.woosan.hr_system.schedule.dao;
-
-public class ClientDAO {
-}
