--- src/main/java/com/woosan/hr_system/schedule/dao/ScheduleDAO.java
@@ -1,4 +1,0 @@
-package com.woosan.hr_system.schedule.dao;
-
-public class ScheduleDAO {
-}
