--- src/main/java/com/woosan/hr_system/schedule/model/Client.java
@@ -1,4 +1,0 @@
-package com.woosan.hr_system.schedule.model;
-
-public class Client {
-}
