--- src/main/java/com/woosan/hr_system/schedule/model/Schedule.java
@@ -1,4 +1,0 @@
-package com.woosan.hr_system.schedule.model;
-
-public class Schedule {
-}
