--- src/main/java/com/woosan/hr_system/schedule/service/ScheduleService.java
@@ -1,4 +1,0 @@
-package com.woosan.hr_system.schedule.service;
-
-public interface ScheduleService {
-}
