--- src/main/java/com/woosan/hr_system/schedule/service/ScheduleServiceImpl.java
@@ -1,6 +1,0 @@
-package com.woosan.hr_system.schedule.service;
-
-import com.woosan.hr_system.schedule.service.ScheduleService;
-
-public class ScheduleServiceImpl implements ScheduleService {
-}
