+++ src/main/resources/static/js/common/layout.js
