+++ src/main/resources/static/js/common/notification.js
