+++ src/main/resources/static/js/report/report.js
