+++ src/main/resources/templates/admin/report/request-edit.html
@@ -8,10 +8,9 @@
 </th:block>
 
 <th:block layout:fragment="script">
-    <!--  <script th:src="@{/js/example.js}"></script> &lt;!&ndash; 외부 JS도 요렇게 &ndash;&gt;-->
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script th:src="@{/js/report.js}"></script>
-    <script th:src="@{/js/employee.js}"></script>
+    <script th:src="@{/js/report/report.js}"></script>
+    <script th:src="@{/js/common/format.js}"></script>
 
     <script>
         $(document).ready(function () {
