+++ src/main/resources/templates/admin/report/request-write.html
@@ -3,12 +3,10 @@
       xmlns:layout="http://www.ultraq.net.nz/web/thymeleaf/layout"
       layout:decorate="~{fragments/layout :: layout(title='보고서 작성 요청')}">
 
-<th:block layout:fragment="css">
-    <!--    <link rel="stylesheet" th:href="@{/css/layout.css}">-->
-</th:block>
+<th:block layout:fragment="css"></th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/report.js}"></script>
+    <script th:src="@{/js/report/report.js}"></script>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script>
         $(document).ready(function () {
