+++ src/main/resources/templates/auth/pwd-management.html
@@ -60,74 +60,14 @@
 
 <th:block layout:fragment="script">
     <script src="https://cdn.jsdelivr.net/npm/zxcvbn@4.4.2/dist/zxcvbn.js"></script>
-    <script th:src="@{/js/pwd.js}"></script>
-    <script th:src="@{/js/employee.js}"></script>
+    <script th:src="@{/js/common/pwd.js}"></script>
+    <script th:src="@{/js/common/format.js}"></script>
     <script>
         document.addEventListener("DOMContentLoaded", function() {
-            calculateDateDifference();
             updateDateTimeElements();
+            calculateDateDifference();
             updateBatteryLevel();
         });
-
-        function updateBatteryLevel() {
-            var strength = parseInt(document.getElementById('password-strength').innerText);
-            var batteryLevel = document.querySelector('.battery-level');
-            var indicatorText = document.querySelector('.battery-indicator');
-
-            // 초기화
-            batteryLevel.innerHTML = '';
-
-            // 배터리 셀을 강도에 맞춰 추가
-            var color;
-            var indicator;
-            var textColor;
-            switch (strength) {
-                case 4:
-                    color = 'linear-gradient(to right, #18A8F1, #5EC5F8)';
-                    indicator = '매우 강함';
-                    textColor = '#18A8F1';
-                    break;
-                case 3:
-                    color = 'linear-gradient(to right, #32C113, #32C113)';
-                    indicator = '강함';
-                    textColor = '#32C113';
-                    break;
-                case 2:
-                    color = 'linear-gradient(to right, #F1CC18, #F1CC18)';
-                    indicator = '보통';
-                    textColor = '#F1CC18';
-                    break;
-                case 1:
-                    color = 'linear-gradient(to right, #F16018, #F16018)';
-                    indicator = '약함';
-                    textColor = '#F16018';
-                    break;
-                default:
-                    color = 'linear-gradient(to right, #F1183C, #F1183C)';
-                    indicator = '매우 약함';
-                    textColor = '#F1183C';
-            }
-
-            // 강도에 따른 셀 개수 및 색상 적용
-            for (var i = 0; i <= strength; i++) {
-                var cell = document.createElement('div');
-                cell.className = 'battery-cell on';
-                cell.style.backgroundImage = color;
-                batteryLevel.appendChild(cell);
-            }
-
-            // 빈 셀 추가
-            for (var i = strength + 1; i < 5; i++) {
-                var cell = document.createElement('div');
-                cell.className = 'battery-cell';
-                cell.style.backgroundColor = '#e6e6e6';
-                batteryLevel.appendChild(cell);
-            }
-
-            // 인디케이터 텍스트 색상 변경
-            indicatorText.innerText = indicator;
-            indicatorText.style.color = textColor;
-        }
     </script>
 </th:block>
