+++ src/main/resources/templates/common/myInfo.html
@@ -8,8 +8,8 @@
 </th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/pwd.js}"/></script>
-    <script th:src="@{/js/employee.js}"></script>
+    <script th:src="@{/js/common/pwd.js}"></script>
+    <script th:src="@{/js/common/format.js}"></script>
     <script>
         document.addEventListener("DOMContentLoaded", function() {
             updateDateTimeElements();
