+++ src/main/resources/templates/employee/list.html
@@ -9,8 +9,8 @@
 </th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/employee.js}"></script>
-    <script th:src="@{/js/employee-detail-edit.js}"></script>
+    <script th:src="@{/js/employee/employee.js}"></script>
+    <script th:src="@{/js/employee/employee-detail-edit.js}"></script>
 </th:block>
 
 <div layout:fragment="~{content}">
