+++ src/main/resources/templates/fragments/layout-with-modal.html
@@ -6,13 +6,12 @@
     <title th:text="${title}">하루하루</title>
 
     <link rel="stylesheet" th:href="@{/css/layout.css}">
-    <script th:src="@{/js/layout.js}"></script>
+    <script th:src="@{/js/common/layout.js}"></script>
 
     <link rel="stylesheet" th:href="@{/css/notification.css}">
-    <script th:src="@{/js/notification.js}"></script>
-    <script>
+    <script th:src="@{/js/common/notification.js}"></script>
 
-    </script>
+    <script th:src="@{/js/common/format.js}"></script>
 
     <link rel="stylesheet" th:href="@{/css/modal.css}">
