+++ src/main/resources/templates/fragments/layout.html
@@ -6,10 +6,10 @@
         <title th:text="${title}">하루하루</title>
 
         <link rel="stylesheet" th:href="@{/css/layout.css}">
-        <script th:src="@{/js/layout.js}"></script>
+        <script th:src="@{/js/common/layout.js}"></script>
 
         <link rel="stylesheet" th:href="@{/css/notification.css}">
-        <script th:src="@{/js/notification.js}"></script>
+        <script th:src="@{/js/common/notification.js}"></script>
 
         <!-- 각 페이지의 script가 작성될 위치 -->
         <th:block layout:fragment="script"></th:block>
