+++ src/main/resources/templates/report/main.html
@@ -9,8 +9,8 @@
 
 <th:block layout:fragment="script">
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script th:src="@{/js/report.js}"></script>
-    <script th:src="@{/js/employee.js}"></script>
+    <script th:src="@{/js/report/report.js}"></script>
+    <script th:src="@{/js/employee/employee.js}"></script>
     <script>
         document.addEventListener("DOMContentLoaded", function () {
             updateDateTimeElements();
