+++ src/main/resources/templates/report/request-list.html
@@ -9,8 +9,8 @@
 </th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/report.js}"></script>
-    <script th:src="@{/js/employee.js}"></script>
+    <script th:src="@{/js/report/report.js}"></script>
+    <script th:src="@{/js/employee/employee.js}"></script>
     <script>
         document.addEventListener("DOMContentLoaded", function () {
             updateDateTimeElements();
