+++ src/main/resources/templates/report/write-from-request.html
@@ -8,8 +8,8 @@
 </th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/report.js}"></script>
-    <script th:src="@{/js/report-file.js}"></script>
+    <script th:src="@{/js/report/report.js}"></script>
+    <script th:src="@{/js/report/report-file.js}"></script>
     <script>
         function submitReportForm(event) {
             // 유효성 검사 실행
