+++ src/main/resources/templates/report/write.html
@@ -8,9 +8,9 @@
 </th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/report.js}"></script>
+    <script th:src="@{/js/report/report.js}"></script>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script th:src="@{/js/report-file.js}"></script>
+    <script th:src="@{/js/report/report-file.js}"></script>
     <script>
         $(document).ready(function () {
             initEventListeners();
